Split Google HTTP API into UA and GA4 specific parts
